move example to scatterplot, rm commented code
