--- test/apps/webgpu-point-cloud/README.md
@@ -1,19 +1,0 @@
-This is a minimal standalone version of the PointCloudLayer example
-on [deck.gl](http://deck.gl) website, ported to WebGPU.
-
-### Usage
-
-Copy the content of this folder to your project. 
-
-```bash
-# install dependencies
-npm install
-# or
-yarn
-# bundle and serve the app with vite
-npm start
-```
-
-### Data Source
-
-Sample data is stored in [deck.gl Example Data](https://github.com/visgl/deck.gl-data/tree/master/examples/line), [Airports data source](http://www.naturalearthdata.com/)
