--- test/apps/webgpu-point-cloud/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "deckgl-examples-point-cloud",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "scripts": {
-    "start": "vite --open",
-    "start-local": "vite --config ../vite.config.local.mjs",
-    "build": "vite build"
-  },
-  "dependencies": {
-    "@loaders.gl/las": "^4.2.0",
-    "@luma.gl/core": "file://../../../../luma.gl/modules/core",
-    "@luma.gl/webgpu": "file://../../../../luma.gl/modules/webgpu",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "deck.gl": "^9.1.3",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "devDependencies": {
-    "typescript": "^5.7.0",
-    "vite": "^6.0.0"
-  }
-}
