--- test/apps/webgpu-point-cloud/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "ES2022",
-    "target": "ES2020",
-    "jsx": "react",
-    "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true
-  }
-}
